move syllabus folder from remove-js-modules
