+++ src/common/components/MarkDownParser/ArticleMarkdown.jsx
@@ -5,7 +5,7 @@ import Toc from './toc';
 import ContentHeading from './ContentHeading';
 import SubTasks from './SubTasks';
 import MarkDownParser from './index';
-import ProjectInstructions from '../../../js_modules/syllabus/ProjectInstructions';
+import ProjectInstructions from '../GuidedExperience/ProjectInstructions';
 
 function ArticleMarkdown({
   content, withToc, frontMatter, titleRightSide, currentTask, currentData,
