+++ src/pages/syllabus/[cohortSlug]/[lesson]/[lessonSlug]/index.jsx
@@ -24,12 +24,12 @@ import ShareButton from '../../../../../common/components/ShareButton';
 import ModalInfo from '../../../../../common/components/ModalInfo';
 import ReactPlayerV2 from '../../../../../common/components/ReactPlayerV2';
 import ScrollTop from '../../../../../common/components/scrollTop';
-import TimelineSidebar from '../../../../../js_modules/syllabus/TimelineSidebar';
-import GuidedExperienceSidebar from '../../../../../js_modules/syllabus/GuidedExperienceSidebar';
-import ExerciseGuidedExperience from '../../../../../js_modules/syllabus/ExerciseGuidedExperience';
-import ProjectBoardGuidedExperience from '../../../../../js_modules/syllabus/ProjectBoardGuidedExperience';
-import SyllabusMarkdownComponent from '../../../../../js_modules/syllabus/SyllabusMarkdownComponent';
-import Topbar from '../../../../../js_modules/syllabus/Topbar';
+import TimelineSidebar from '../../../../../common/components/GuidedExperience/TimelineSidebar';
+import GuidedExperienceSidebar from '../../../../../common/components/GuidedExperience/GuidedExperienceSidebar';
+import ExerciseGuidedExperience from '../../../../../common/components/GuidedExperience/ExerciseGuidedExperience';
+import ProjectBoardGuidedExperience from '../../../../../common/components/GuidedExperience/ProjectBoardGuidedExperience';
+import SyllabusMarkdownComponent from '../../../../../common/components/GuidedExperience/SyllabusMarkdownComponent';
+import Topbar from '../../../../../common/components/GuidedExperience/Topbar';
 import bc from '../../../../../common/services/breathecode';
 import useCohortHandler from '../../../../../common/hooks/useCohortHandler';
 import SimpleModal from '../../../../../common/components/SimpleModal';
